+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added envied config for managing environment variables.
+- Added API client starter setup (Dio + Retrofit).
 
 ### Updated